Move channel oidc tests
